Rename all components in containers/ replacing View -> Container.
